Switch translation tests to 'it' as 'es' shows strange results. (#119)
Switch the translation tests to 'it' as the up to now used 'es' now shows the English translation 'Hello' instead of the expected 'Hola'. No idea, why the result now is different than before. This may get fixed in the future, then we can switch back again. For the time being, the translation to 'it' shows the expected translation 'Ciao' and we use that for the tests.